requirements.txt
pillow>=9.0.0
pandas>=1.3.0
numpy>=1.21.0
asammdf>=7.0.0
openpyxl>=3.0.0
xlrd>=2.0.0 